Make the Carob database

Make the Carob database.

run_carob runs a single Carob script for a dataset. Carob scripts download data to "<path>/data/raw", re-organize the data, check the variable names and controlled vocabulary used, check data types, ranges, and location of the coordinates, and write clean data to "carob/data/clean"

process_carob runs all the Carob scripts.

compile_carob combineds all the standardized data by "group" and saves these to "<path>/data/compiled")

make_carob is a convenience function that first runs process_carob and then runs compile_carob

run_carob(cleanuri, path, group="", quiet=FALSE)
process_carob(path, group="", quiet=FALSE, check=NULL, cache=TRUE, purge=FALSE)
compile_carob(path, group="", split_license=FALSE, zip=FALSE, excel=FALSE, cache=TRUE)
make_carob(path, group="", quiet=FALSE, check="all", report=FALSE, combine=FALSE, cache=TRUE, purge=FALSE, ...)

Arguments

cleanuri

character. A "clean" URI made with simpleURI

path

character. Local path to the Carob repo

group

character. group name

quiet

logical. Be quiet?

check

character. "none" to do no checking; "nogeo" to not check coordinates. "all" and all other values to do all checks

combine

logical. Should all data also be compiled into one file?

cache

logical. If TRUE, datasets are only recompiled if the R script has changed after the last compilation

report

logical. Should a html report be created for the datasets?

split_license

logical. Make a separate aggregation for data with a CC license?

excel

logical. Also write the results into an excel file?

zip

logical. zip the aggregated data?

purge

logical. If TRUE, datasets with the wrong version number are purged (removed)

...

additional arguments passed to compile_carob

Value

none. Used for the side effect of creating files.