Write Carob dataset files

Write Carob dataset files. This calls check_terms

write_files(path=NULL, metadata, wide, long=NULL, wth=NULL, var_meta=NULL, options=NULL)

Arguments

path

character. NULL, or the path to the local Carob repo

metadata

data.frame. One row with dataset level descriptive variables

wide

data.frame. Dataset with records in the "wide" format (one row for each observational unit)

long

data.frame. Dataset with records in the "long" format to capture multiple observations over time or depth for a single variable and observational unit

wth

data.frame with weather data

var_meta

data.frame. Metadata associated with speficic variables (e.g. extraction method used in chemical soil analysis)

options

list

Value

if path=NULL a list of data.frames (metadata and records). Otherwise a logical value indicating if the standardized dataset was succesfully written to the Carob repo.