Download a data set

Download the files associated with a URI.

get_data(uri, path, group, files=NULL, cache=TRUE, recursive=FALSE, filter=TRUE, auto_json_bundle=TRUE, protocol="")

Arguments

uri

character. The URI of a dataset

path

character. The path to the Carob repo

group

character. group name

files

character. If not NULL these files are copied to the raw data folder; and uri is not used to get the data. This can be useful for datasets that do not have a URI or are not hosted on a website with a standard protocol for data retrieval such as dataverse or CKAN. It can also be useful to process local files before they are published. Files can be either *all* have a http address, or *all* on the local file system

cache

logical. if TRUE the data are only downloaded again if they are not locally available

recursive

logical. if TRUE also return filenames in sub-directories

filter

logical. if TRUE files that probably do not have the primary data are filtered out (files ending on .pdf, .doc(x), .json)

auto_json_bundle

logical. if TRUE (default), when the shallow file list has no tabular files but the raw folder has subdirectories or .json files other than the dataset descriptor (<dataset_id>.json) and metadata.json, get_data re-lists with recursive=TRUE and skips filtering so those JSON paths stay in the returned vector (Dataverse-style nested Dataset/*.json bundles).

protocol

character; use "LSMS" for LSMS (usr_pwd) downloads

Value

character